requirements.txt
asgiref==3.5.1
Django==4.0.4
django-appconf==1.0.5
django-imagekit==4.1.0
pilkit==2.0
Pillow==9.1.0
psycopg2-binary==2.9.3
python-decouple==3.6
six==1.16.0
sqlparse==0.4.2
whitenoise==6.2.0
